refactor Operand and Option to use Argument class methods for value
Unfortunately this leads to a wrong message when the value is invalid - both are
now showing 'Operand %s has an invalid value'. But this will be solved later in
this branch.
